Moves resolvers to resolvers directory
