+++ taxi-ui/src/app/resolvers/trip-detail.resolver.spec.ts
