+++ taxi-ui/src/app/resolvers/trip-list.resolver.spec.ts
